[compiler][llvm] Emit code using 64bit integer

+++ samlang-demo/src/__test__/index.test.ts
@@ -13,20 +13,20 @@ function _Demo_Main_main() {
   return 0;
 }
 `,
-      llvmString: `declare i32* @_builtin_malloc(i32) nounwind
-declare i32 @__Builtins_println(i32*) nounwind
-declare i32 @__Builtins_panic(i32*) nounwind
-declare i32* @__Builtins_intToString(i32) nounwind
-declare i32 @__Builtins_stringToInt(i32*) nounwind
-declare i32* @_builtin_stringConcat(i32*, i32*) nounwind
+      llvmString: `declare i64* @_builtin_malloc(i64) nounwind
+declare i64 @__Builtins_println(i64*) nounwind
+declare i64 @__Builtins_panic(i64*) nounwind
+declare i64* @__Builtins_intToString(i64) nounwind
+declare i64 @__Builtins_stringToInt(i64*) nounwind
+declare i64* @_builtin_stringConcat(i64*, i64*) nounwind
 
 ; @GLOBAL_STRING_0 = 'hello world'
-@GLOBAL_STRING_0 = private unnamed_addr constant [12 x i32] [i32 11, i32 104, i32 101, i32 108, i32 108, i32 111, i32 32, i32 119, i32 111, i32 114, i32 108, i32 100], align 8
-define i32 @_Demo_Main_main() local_unnamed_addr nounwind {
+@GLOBAL_STRING_0 = private unnamed_addr constant [12 x i64] [i64 11, i64 104, i64 101, i64 108, i64 108, i64 111, i64 32, i64 119, i64 111, i64 114, i64 108, i64 100], align 8
+define i64 @_Demo_Main_main() local_unnamed_addr nounwind {
 l0_start:
-  %_temp_0_string_name_cast = bitcast [12 x i32]* @GLOBAL_STRING_0 to i32*
-  call i32 @__Builtins_println(i32* %_temp_0_string_name_cast) nounwind
-  ret i32 0
+  %_temp_0_string_name_cast = bitcast [12 x i64]* @GLOBAL_STRING_0 to i64*
+  call i64 @__Builtins_println(i64* %_temp_0_string_name_cast) nounwind
+  ret i64 0
 }`,
       errors: [],
     });
@@ -37,12 +37,12 @@ l0_start:
       prettyPrintedProgram: 'class Main {  }\n',
       interpreterPrinted: '',
       jsString: '',
-      llvmString: `declare i32* @_builtin_malloc(i32) nounwind
-declare i32 @__Builtins_println(i32*) nounwind
-declare i32 @__Builtins_panic(i32*) nounwind
-declare i32* @__Builtins_intToString(i32) nounwind
-declare i32 @__Builtins_stringToInt(i32*) nounwind
-declare i32* @_builtin_stringConcat(i32*, i32*) nounwind
+      llvmString: `declare i64* @_builtin_malloc(i64) nounwind
+declare i64 @__Builtins_println(i64*) nounwind
+declare i64 @__Builtins_panic(i64*) nounwind
+declare i64* @__Builtins_intToString(i64) nounwind
+declare i64 @__Builtins_stringToInt(i64*) nounwind
+declare i64* @_builtin_stringConcat(i64*, i64*) nounwind
 `,
       assemblyString: undefined,
       errors: [],
